Crate antimony_spawn

This application is an implementation of flatpak-spawn for Antimony. Some of its arguments are no-ops, and only exist for a compatible interface, but it’s largely feature complete. There are two important distinctions:

The dichotomy between –sandbox and –host are vastly different. When running in –sandbox, the requested command is run in a new Antimony sandbox, constructed from the command line arguments. Because the environment antimony-spawn runs in (Antimony itself) does not have an Antimony installation, we have no features or profiles; therefore, we pass the sandbox’s lib folder directly–which has already been reduced to only needing what the main application and its dependencies require.

–host does not run a command on the host. It just runs the command directly in the sandbox. Allowing execution on the host, even mediated, is a tremendous vulnerability, and trying to implement it would require significant effort.

–watch-bus is a no-op: antimony-spawn will either exit if the child it runs does, or if it is hit with SIGTERM or SIGINT.

Structs

Cli

Functions

main 🔒